VT module fixes from Jack; Updated build scripts.

+++ src/vt/l_tr1.cpp
@@ -766,7 +766,7 @@ void TR_Level::read_tr_level(SDL_RWops * const src, bool demo_or_ub)
         this->sound_details[i].num_samples_and_flags_1 = read_bitu8(src);
         this->sound_details[i].flags_2 = read_bitu8(src);
         this->sound_details[i].sound_range = TR_AUDIO_DEFAULT_RANGE;
-        this->sound_details[i].pitch = TR_AUDIO_DEFAULT_PITCH;
+        this->sound_details[i].pitch = (int16_t)TR_AUDIO_DEFAULT_PITCH;
     }
 
     // LOAD SAMPLES
